+++ packages/components-js/tests/a11y/specs/axe-core/common/components.a11y.ts
@@ -56,15 +56,7 @@ components.forEach((component) => {
               contentType: 'application/json',
             });
 
-            if (component === 'canvas' && viewportWidth === viewportWidthM) {
-              // TODO: fix asap
-              expect(accessibilityScanResults.violations.length).toBe(3);
-            } else if (component === 'canvas' && viewportWidth === viewportWidthXXS) {
-              // TODO: fix asap
-              expect(accessibilityScanResults.violations.length).toBe(2);
-            } else {
-              expect(accessibilityScanResults.violations.length).toBe(0);
-            }
+            expect(accessibilityScanResults.violations.length).toBe(0);
           });
         });
       } else {